requirements.txt
fastapi
uvicorn
openai
jinja2
python-dotenv
aiosqlite

